% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_clean_data.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Clean Data}
\source{
The raw dataset is obtained from National Geophysical 
  Data Center / World Data Service (NGDC/WDS): Significant 
  Earthquake Database. National Geophysical Data Center, NOAA. 
  (\href{http://dx.doi.org/10.7289/V5TD9V7K}{doi:10.7289/V5TD9V7K})
}
\usage{
eq_clean_data(file)
}
\arguments{
\item{file}{A character string representing the dataset file
path and file name}
}
\value{
This function returns a cleaned dataframe, from \code{YEAR = 1900} 
onwards, with the following 8 fields:
\itemize{
 \item{DATE}
 \item{LATITUDE}
 \item{LONGITUDE}
 \item{COUNTRY}
 \item{LOCATION}
 \item{MAGNITUDE}
 \item{INTENSITY}
 \item{DEATH_COUNT}
}
}
\description{
This function takes the raw dataset and returns 
it in a readable format. The function also omits records with
\code{NA} values for \code{YEAR}, \code{MONTH}, \code{DAY}, 
\code{EQ_PRIMARY} and \code{INTENSITY}. If \code{TOTAL_DEATHS} 
is \code{NA}, then the function will return \code{DEATH_COUNT = 0}. 
If the file does not exist, then the function will print an error 
message and return a \code{NULL} object.
}
\examples{
\dontrun{eq_clean_data("./inst/extdata/dataset.txt")}
}
