% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_create_label.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Create Map Label}
\usage{
eq_create_label(filtered_data)
}
\arguments{
\item{filtered_data}{A subset dataframe from the \code{eq_clean_data}
function for the country and period of interest}
}
\value{
A list of character type representing the description for each
  earthquake, including its location, magnitude, and total deaths
}
\description{
This function takes a filted dataframe object and create
\code{html} format data labels for each earthquake point.
}
\examples{
\dontrun{eq_create_label(.)}
}
