% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_location_clean.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Clean Data (Location)}
\usage{
eq_location_clean(clean_data)
}
\arguments{
\item{clean_data}{A dataframe object returned by the 
\code{eq_clean_data} function}
}
\value{
This function returns a cleaned dataframe, with the 
following 8 fields:
\itemize{
 \item{DATE}
 \item{LATITUDE}
 \item{LONGITUDE}
 \item{COUNTRY}
 \item{LOCATION}
 \item{MAGNITUDE}
 \item{INTENSITY}
 \item{DEATH_COUNT}
}
}
\description{
This function takes a dataframe object from the
\code{eq_clean_data} function and modifies the \code{LOCATION}
by removing the \code{COUNTRY} substring from the \code{LOCATION}
value. If the object is not a dataframe or has misnamed columns
for \code{COUNTRY} or \code{LOCATION}, then the function will
print an error message and return a \code{NULL} object.
}
\examples{
\dontrun{eq_location_clean(clean_data)}
}
