% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_map.R
\name{eq_map}
\alias{eq_map}
\title{Create Map}
\usage{
eq_map(filtered_data, annot_col = "DATE")
}
\arguments{
\item{filtered_data}{A subset dataframe from the \code{eq_clean_data}
function for the country and period of interest}

\item{annot_col}{A character string representing the column name from
the dataset from which the label of earthquake should be obtained}
}
\value{
A leaflet object representing the map of earthquake points
}
\description{
This function takes a filted dataframe object and create
a map of the earthquakes with a popup label describing the earthquakes.
}
\examples{
\dontrun{readr::read_delim("./inst/extdata/dataset.txt", delim = "\\t") \%>\% 
  eq_clean_data() \%>\% 
  eq_location_clean() \%>\%
  filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000) \%>\% 
  mutate(popup_text = eq_create_label(.)) \%>\% 
  eq_map(annot_col = "popup_text")}
}
